clone: preserve range, progress, and sync semantics v5
Block cloning spans common DMU reservation, regular-file and zvol
callers, BRT dirty records, page-cache repair, partial progress, and
synchronous durability. Treating those as independent fragments left
holes: reservations could miss partial edge blocks, a NOFILL redirty
could retain clone state, a late zvol error could hide a committed
prefix, wide repair lengths could truncate, and O_SYNC could return
before the cloned prefix reached the ZIL.

Count the actual touched L0 blocks, cancel pending clone state before
NOFILL redirty, report committed zvol progress, keep page-cache repair
lengths 64-bit, and commit synchronous file clones before return.

Why carried: This is generic block-clone correctness across regular
  files, zvols, transaction holds, page cache, and durability.

Design: A clone reserves every affected block, reports the exact
  committed prefix, clears superseded BRT state, repairs the full cache
  range, and honors the destination synchronous contract. OpenZFS 2.3.8
  independently supplies the encrypted cross-key rejection, so that
  formerly folded fragment is no longer part of this downstream diff.

Provenance: Original downstream work by Pavel Snajdr, consolidating the
  zvol progress, DMU hold, synchronous clone, NOFILL state, and
  page-cache repair lineages. The earlier zvol key-policy fragment was
  dropped in favor of OpenZFS 2.3.8 commit
  e2705f63de5a87b0bc048ac3bba2df5bf319c595.

Patch development history:
- v1: Rejected encrypted zvol clones across different master keys and
  tracked partial zvol progress.
- v2: Counted clone transaction holds from the actual touched blocks,
  including partial edges.
- v3: Propagated O_SYNC/O_DSYNC into regular-file clone completion and
  committed the cloned prefix through ZIL.
- v4: Folded NOFILL clone-state cancellation and 64-bit page-cache
  repair into one end-to-end clone contract.
- v5: Ported onto OpenZFS 2.3.8 and removed the now-upstream
  encrypted-key fragment while retaining the remaining range, progress,
  state, and durability corrections.

Validation: git diff --cached --check and scripts/cstyle.pl -cpP passed
  for all changed C and header files before commit. No build or runtime
  test was run during this history reconstruction and OpenZFS 2.3.8
  port.

+++ module/os/linux/zfs/zfs_vnops_os.c
@@ -484,7 +484,7 @@ zfs_read_mapped_range(znode_t *zp, uint64_t start, uint64_t len, void *buf,
  * pages with the contents of the coresponding dmu buffer.
  */
 void
-update_pages(znode_t *zp, int64_t start, int len, objset_t *os)
+update_pages(znode_t *zp, int64_t start, uint64_t len, objset_t *os)
 {
 	struct address_space *mp = ZTOI(zp)->i_mapping;
 	int64_t off = start & (PAGE_SIZE - 1);
